comparison with creator

+++ routes/gallery/deleteGallery.js
@@ -10,15 +10,24 @@ const Gallery = require('../../models/gallery');
 
 const deleteGallery = async function(req, res){
   try{
-    var gallery = await Gallery.deleteOne({_id: req.params.galleryId});
-    if(gallery && gallery.deletedCount > 0){
-      return res.status(200).send({
-        message: 'Gallery deleted successfully.',
+    var result = await Gallery.findById(req.params.galleryId);
+    var owner = req.user.email;
+    if(owner === result.creator){
+      var gallery = await Gallery.deleteOne({_id: req.params.galleryId});
+      if(gallery && gallery.deletedCount > 0){
+        return res.status(200).send({
+          message: 'Gallery deleted successfully.',
+        });
+      }
+      return res.status(404).send({
+        message: 'Gallery not found.',
       });
     }
-    return res.status(404).send({
-      message: 'Gallery not found.',
-    });
+    else {
+      return res.status(403).send({
+        message: 'No permission deleting the gallery project.',
+      });
+    }
   }
   catch(err){
     return res.status(500).send(err);
